requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==4.0.3
attrs==25.1.0
bidict==0.23.1
blinker==1.9.0
cachelib==0.13.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dataclasses-json==0.5.14
dnspython==1.16.0
exceptiongroup==1.2.2
Flask==3.1.0
Flask-Cors==3.0.10
Flask-PyMongo==3.0.1
Flask-Session==0.6.0
Flask-SocketIO==5.5.1
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
ibm-cos-sdk==2.13.6
ibm-cos-sdk-core==2.13.6
ibm-cos-sdk-s3transfer==2.13.6
ibm_watsonx_ai==1.2.8
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
jsonify==0.5
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.19
langchain-core==0.3.37
langchain-ibm==0.3.6
langchain-text-splitters==0.3.6
langsmith==0.3.10
lomond==0.3.3
MarkupSafe==3.0.2
marshmallow==3.26.1
msgspec==0.19.0
multidict==6.1.0
mypy-extensions==1.0.0
numexpr==2.10.2
numpy==1.26.4
openapi-schema-pydantic==1.2.4
orjson==3.10.15
packaging==24.2
pandas==2.1.4
propcache==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==3.12.0
python-dateutil==2.9.0.post0
python-dotenv==0.19.0
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.2
requests-toolbelt==1.0.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
tabulate==0.9.0
tenacity==8.5.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
